// ui/src/app/Modules/employee/models/EmpBasicDetails.model.ts
export interface EmpBasicDetails {
    empId: number;
    departmentId: number;
    departmentName: string;
    fullName: string;
    gender: string;
    profileImage: string;
    empCode: string;
    roleId: number;
    roleName: string;
    emailId: string
    reportingManager: string;
    teamId: number;
    teamName: string;
    empType: string;
    actualEmployementDate: number;
    pastExperience: string;
    dateOfJoining: number;
    permanentAddress: string;
    presentAddress: string;
    phoneNumber: string;
    emergencyNumber: string;
    panNo: string;
    aadharNo: string;
    dateOfBirth: number;
    salaryID: number;
    ctc: string;
    basicSalary: string;
    hra: string;
    pf: string;
    other: string;
    raise: string;
    payrollCode: string;
    ratingId: number;
    rating: number;
    createdDate: number;
    modifiedDate: number;
    isActive: boolean;
}